requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.1
django-environ==0.10.0
idna==3.4
marshmallow==3.17.0
packaging==23.1
Pillow==9.5.0
protobuf==3.20.3
requests==2.30.0
sqlparse==0.4.4
transbank-sdk==4.0.0
urllib3==2.0.2
